+++ CHANGELOG.md
@@ -8,3 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - Initial release with:
   - Tools/01 Build Spritesheet From Dumps
+  - Tools/03 Build Bundles
+  - Tools/04 Fix Dump Files (Sprites)
+  - Tools/05 Assign Sprites to SpriteAtlas (Atlas)
+  - Tools/06 Fix Dump Files (Atlas)
+  - Tools/07 Set Audio
+  - Tools/08 Replace Audio